+++ th2_data_services/data_source/lwdp/source_api/http.py
@@ -60,21 +60,49 @@ class HTTPAPI(IHTTPSourceAPI):
         """REST-API `book/{bookID}/event/scopes` call returns a list of scopes in book named bookID."""
         return self.__encode_url(f"{self._url}/book/{book_id}/event/scopes")
 
-    def get_url_get_message_aliases(self, book_id: str) -> str:
-        """REST-API `book/{bookID}/message/aliases` call returns a list of message aliases in book named bookID."""
-        return self.__encode_url(f"{self._url}/book/{book_id}/message/aliases")
+    def get_url_get_message_aliases(
+        self,
+        book_id: str,
+        start_timestamp: int = None,
+        end_timestamp: int = None,
+        chunked_size: int = None,
+    ) -> str:
+        """REST-API `book/{bookID}/message/aliases/sse` call creates an sse channel of message aliases in book named bookID."""
+        url = f"{self._url}/book/{book_id}/message/aliases/sse?"
 
-    def get_url_get_message_groups(self, book_id: str) -> str:
-        """REST-API `book/{bookID}/message/groups` call returns a list of message groups in book named bookID."""
-        return self.__encode_url(f"{self._url}/book/{book_id}/message/groups")
+        kwargs = {
+            "startTimestamp": start_timestamp,
+            "endTimestamp": end_timestamp,
+            "chunkedSize": chunked_size,
+        }
 
+        url += "&".join(f"{k}={v}" for k, v in kwargs.items() if v is not None)
+        return self.__encode_url(url)
+
+    def get_url_get_message_groups(
+        self,
+        book_id: str,
+        start_timestamp: int = None,
+        end_timestamp: int = None,
+        chunked_size: int = None,
+    ) -> str:
+        """REST-API `book/{bookID}/message/groups/sse` call creates an sse channel of message groups in book named bookID."""
+        url = f"{self._url}/book/{book_id}/message/groups/sse?"
+
+        kwargs = {
+            "startTimestamp": start_timestamp,
+            "endTimestamp": end_timestamp,
+            "chunkedSize": chunked_size,
+        }
+
+        url += "&".join(f"{k}={v}" for k, v in kwargs.items() if v is not None)
+        return self.__encode_url(url)
+    
     def get_url_find_event_by_id(self, event_id: str) -> str:
         """REST-API `event` call returns a single event with the specified id."""
         return self.__encode_url(f"{self._url}/event/{event_id}")
 
-    def get_url_find_message_by_id(
-        self, message_id: str, only_raw: bool = False
-    ) -> str:
+    def get_url_find_message_by_id(self, message_id: str, only_raw: bool = False) -> str:
         """REST-API `message` call returns a single      message with the specified id."""
         return self.__encode_url(f"{self._url}/message/{message_id}?onlyRaw={only_raw}")